Removed CLI util start-e2e
